refactor: Car 객체 getter 접근 수정
