+++ src/test/java/racingcar/model/CarTest.java
@@ -1,20 +1,17 @@
-package racingcar;
+package racingcar.model;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
+import racingcar.model.Car;
+
 public class CarTest {
-	@Test
-	void 자동차_생성() {
-		Car car = new Car("소주캉");
-		assertThat(car.getName()).isEqualTo("소주캉");
-	}
 
 	@Test
 	void 자동차_초기위치() {
 		Car car = new Car("소주캉");
-		assertThat(car.getPosition()).isEqualTo(0);
+		assertThat(car.isSamePosition(0)).isTrue();
 	}
 
 	@Test
@@ -40,7 +37,9 @@ public class CarTest {
 	void 자동차_이동() {
 		Car car = new Car("범고래");
 		car.move(4);
-		assertThat(car.getPosition()).isEqualTo(1);
+		assertThat(car.isSamePosition(1)).isTrue();
 	}
+
+
 
 }
